Move entrypoint file for better handling of dependencies
